llm_demo/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0
torch>=2.0.0
torchvision>=0.15.0

